Moved wagon wheel function from batter page to its own file
